Fixed Range policy tests | Added Chunk element policy tests
