+++ include/compute/fields/simulation_field.tpp
@@ -25,8 +25,6 @@ int compute_nglob(const ViewType index_mapping) {
     }
   }
 
-  std::cout << "nglob: " << nglob << std::endl;
-
   return nglob + 1;
 }
 } // namespace
