+++ tests/unit-tests/CMakeLists.txt
@@ -215,12 +215,12 @@ target_link_libraries(
 )
 
 add_executable(
-  range_policy
-  policies/RangePolicy.cpp
+  policies
+  policies/policies.cpp
 )
 
 target_link_libraries(
-  range_policy
+  policies
   mesh
   source_class
   receiver_class
@@ -372,7 +372,7 @@ if (NOT MPI_PARALLEL)
   # # gtest_discover_tests(compute_acoustic_tests)
   # gtest_discover_tests(compute_coupled_interfaces_tests)
   gtest_discover_tests(compute_tests)
-  gtest_discover_tests(range_policy)
+  gtest_discover_tests(policies)
   gtest_discover_tests(locate_point)
   gtest_discover_tests(interpolate_function)
   gtest_discover_tests(rmass_inverse_tests)
