+++ tests/unit-tests/policies/test_config.yaml
@@ -1,46 +1,73 @@
 Tests:
   # Test 1:
-  - name : "Test 1: Simple mesh with flat topography"
+  - name : "Test 1: Multiple of simd_size"
     description: >
-      Testing mesh reader on a simple mesh with flat topography.
+      Testing range policy with a range that is a multiple of simd_size.
     config:
       nproc : 1
-    databases:
-      - processor : 0
-        filename : "../../../tests/unit-tests/data/mesh/simple_mesh_flat_topography/database.bin"
+    parameters:
+      nglob : 80000
+      nspec : 8000
   # Test 2:
-  - name : "Test 2: Simple mesh with curved topography"
+  - name : "Test 2: N * simd_size + 1"
     description: >
-      Testing mesh reader on a simple mesh with curved topography.
+      Testing range policy with a range that is N * simd_size + 1.
     config:
       nproc : 1
-    databases:
-      - processor : 0
-        filename : "../../../tests/unit-tests/data/mesh/simple_mesh_curved_topography/database.bin"
+    parameters:
+      nglob : 80001
+      nspec : 8001
   # Test 3:
-  - name : "Test 3: Simple mesh with flat ocean bottom"
+  - name : "Test 3: N * simd_size + 2"
     description: >
-      Testing mesh reader on a simple mesh with flat ocean bottom. Fluid-solid interface is at the ocean bottom.
+      Testing range policy with a range that is N * simd_size + 2.
     config:
       nproc : 1
-    databases:
-      - processor : 0
-        filename : "../../../tests/unit-tests/data/mesh/fluid_solid_mesh_flat_ocean_bottom/database.bin"
+    parameters:
+      nglob : 80002
+      nspec : 8002
   # Test 4:
-  - name : "Test 4: Simple mesh with curved ocean bottom"
+  - name : "Test 4: N * simd_size + 3"
     description: >
-      Testing mesh reader on a simple mesh with curved ocean bottom. Fluid-solid interface is at the ocean bottom.
+      Testing range policy with a range that is N * simd_size + 3.
     config:
       nproc : 1
-    databases:
-      - processor : 0
-        filename : "../../../tests/unit-tests/data/mesh/fluid_solid_mesh_curved_ocean_bottom/database.bin"
+    parameters:
+      nglob : 80003
+      nspec : 8003
   # Test 5:
-  - name : "Test 5: Gmesh Example"
+  - name : "Test 5: N * simd_size + 4"
     description: >
-      Testing mesh reader on a Gmesh example mesh.
+      Testing range policy with a range that is N * simd_size + 4.
     config:
       nproc : 1
-    databases:
-      - processor : 0
-        filename : "../../../tests/unit-tests/data/mesh/Gmesh_Example_Stacey/database.bin"
+    parameters:
+      nglob : 80004
+      nspec : 8004
+  # Test 6:
+  - name : "Test 6: N * simd_size + 5"
+    description: >
+      Testing range policy with a range that is N * simd_size + 5.
+    config:
+      nproc : 1
+    parameters:
+      nglob : 80005
+      nspec : 8005
+  # Test 7:
+  - name : "Test 7: N * simd_size + 6"
+    description: >
+      Testing range policy with a range that is N * simd_size + 6.
+    config:
+      nproc : 1
+    parameters:
+      nglob : 80006
+      nspec : 8006
+  # Test 8:
+  - name : "Test 8: N * simd_size + 7"
+    description: >
+      Testing range policy with a range that is N * simd_size + 7.
+    config:
+      nproc : 1
+    parameters:
+      nglob : 80007
+      nspec : 8007
